requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
#
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via httpx
asgiref==3.9.1
    # via
    #   django
    #   django-allauth
    #   django-cors-headers
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
authlib==1.6.4
    # via safety
bandit==1.8.6
    # via -r requirements-dev.in
billiard==4.2.2
    # via celery
brotli==1.1.0
    # via whitenoise
build==1.3.0
    # via pip-tools
celery==5.5.3
    # via
    #   -r requirements.in
    #   django-celery-beat
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==2.0.0
    # via cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   nltk
    #   pip-tools
    #   safety
    #   typer
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
coverage[toml]==7.10.6
    # via pytest-cov
cron-descriptor==2.0.6
    # via django-celery-beat
cryptography==46.0.1
    # via authlib
distlib==0.4.0
    # via virtualenv
dj-database-url==3.0.1
    # via -r requirements.in
dj-rest-auth==7.0.1
    # via -r requirements.in
django==5.2
    # via
    #   -r requirements.in
    #   dj-database-url
    #   dj-rest-auth
    #   django-allauth
    #   django-celery-beat
    #   django-cors-headers
    #   django-filter
    #   django-prometheus
    #   django-redis
    #   django-storages
    #   django-stubs
    #   django-stubs-ext
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   sentry-sdk
django-allauth==65.11.2
    # via -r requirements.in
django-celery-beat==2.8.1
    # via -r requirements.in
django-cors-headers==4.9.0
    # via -r requirements.in
djangorestframework-stubs==3.16.0
    # via -r requirements-dev.in
django-filter==25.2
    # via -r requirements.in
django-prometheus==2.4.1
    # via -r requirements.in
django-redis==6.0.0
    # via -r requirements.in
django-storages==1.14.6
    # via -r requirements.in
django-stubs==5.2.5
    # via -r requirements-dev.in
django-stubs-ext==5.2.5
    # via django-stubs
django-timezone-field==7.1
    # via django-celery-beat
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
dparse==0.6.4
    # via
    #   safety
    #   safety-schemas
drf-spectacular==0.28.0
    # via -r requirements.in
factory-boy==3.3.3
    # via -r requirements-dev.in
model-bakery==1.20.5
    # via -r requirements-dev.in
faker==37.8.0
    # via factory-boy
filelock==3.19.1
    # via
    #   safety
    #   virtualenv
gunicorn==23.0.0
    # via -r requirements.in
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via safety
identify==2.6.14
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
inflection==0.5.1
    # via drf-spectacular
iniconfig==2.1.0
    # via pytest
jinja2==3.1.6
    # via safety
joblib==1.5.2
    # via nltk
jsonschema==4.25.1
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
kombu==5.5.4
    # via celery
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
marshmallow==4.0.1
    # via safety
mdurl==0.1.2
    # via markdown-it-py
mypy==1.18.2
    # via -r requirements-dev.in
mypy-extensions==1.1.0
    # via mypy
nltk==3.9.1
    # via safety
nodeenv==1.9.1
    # via pre-commit
packaging==25.0
    # via
    #   build
    #   dparse
    #   gunicorn
    #   kombu
    #   pytest
    #   safety
    #   safety-schemas
pathspec==0.12.1
    # via mypy
pillow==11.3.0
    # via -r requirements.in
pip-tools==7.5.1
    # via -r requirements-dev.in
platformdirs==4.4.0
    # via virtualenv
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pre-commit==4.3.0
    # via -r requirements-dev.in
prometheus-client==0.23.1
    # via django-prometheus
prompt-toolkit==3.0.52
    # via click-repl
psutil==7.1.0
    # via
    #   -r requirements.in
    #   safety
psycopg2-binary==2.9.10
    # via -r requirements.in
pybreaker==1.2.0
    # via -r requirements.in
pycparser==2.23
    # via cffi
pydantic==2.9.2
    # via
    #   safety
    #   safety-schemas
pydantic-core==2.23.4
    # via pydantic
pygments==2.19.2
    # via
    #   pytest
    #   rich
pyjwt==2.10.1
    # via djangorestframework-simplejwt
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.4.2
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-django
    #   pytest-mock
pytest-cov==7.0.0
    # via -r requirements-dev.in
pytest-django==4.11.1
    # via -r requirements-dev.in
pytest-mock==3.15.1
    # via -r requirements-dev.in
python-crontab==3.3.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via celery
python-decouple==3.8
    # via -r requirements.in
python-json-logger==4.0.0
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   bandit
    #   drf-spectacular
    #   pre-commit
redis==6.4.0
    # via
    #   -r requirements.in
    #   django-redis
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.9.18
    # via nltk
requests==2.32.5
    # via
    #   -r requirements.in
    #   safety
rich==14.1.0
    # via
    #   bandit
    #   typer
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.15
    # via
    #   safety
    #   safety-schemas
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
ruff==0.13.1
    # via -r requirements-dev.in
safety==3.6.1
    # via -r requirements-dev.in
safety-schemas==0.0.14
    # via safety
sentry-sdk[django]==2.41.0
    # via -r requirements.in
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlparse==0.5.3
    # via django
stevedore==5.5.0
    # via bandit
structlog==25.4.0
    # via -r requirements.in
tenacity==9.1.2
    # via safety
tomlkit==0.13.3
    # via safety
tqdm==4.67.1
    # via nltk
typer==0.19.1
    # via safety
types-pyyaml==6.0.12.20250915
    # via django-stubs
typing-extensions==4.15.0
    # via
    #   anyio
    #   cron-descriptor
    #   django-stubs
    #   django-stubs-ext
    #   mypy
    #   pydantic
    #   pydantic-core
    #   referencing
    #   safety
    #   safety-schemas
    #   typer
tzdata==2025.2
    # via
    #   django-celery-beat
    #   faker
    #   kombu
uritemplate==4.2.0
    # via drf-spectacular
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==20.34.0
    # via pre-commit
wcwidth==0.2.13
    # via prompt-toolkit
wheel==0.45.1
    # via pip-tools
whitenoise[brotli]==6.11.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
